Limit chdir over FTP
Rather than resetting to the root upload directory and testing
every directory for existance for every file, if the next file
is in the same directory simply upload immediately.

I more rigorous solution to this performance issue would be to
keep a list of every existing directory, rather than going to
the server each tiem. But since Git appears to return the files
in alphabetical order, and therefore grouped by directory. This
fix should give almost a good performance benefits.