Make it more difficult to deploy a non-fast-forward commit
As well as the previous commit to prevent committing the wrong
branch, it is also possible in a multi-person environment that
you don't have the most recent commits, and therefore deploying
from the tip of your branch would be moving backwards in time.

This commit make Gitriol check that the commit you are deploying
is a fast-forward from the previous deployment.