Checkout commit being deployed if not current
This was a big oversight on my part. I forgot to actually
checkout the commit that is being deployed, so the files in the
current working directory would always get uploaded.

A nice side-effect is that we now check that the working
directory is clean before uploading, since you don't want to
upload any local changes that are not in the repository.